chatbot2/requirements.txt
fastapi==0.114.0
langchain
langserve==0.2.3
uvicorn==0.23.2
langchain-cli==0.0.30
openai==1.44.1
langchain-community
langchain-core
langchain-google-community
firebase
python-dotenv